fix(@angular/cli): Move update check to inProject command logic
fixes #10145
